moving MoEGateConfig to tt/moe
